Revert "Почти закончил PizzaDetailScreen, пока убрал логику с CatalogScreen"
This reverts commit 6319e1eb58145baa4a389a20b35fdf07ff7d2c12.

+++ app/src/main/java/com/example/pizza_shift_2025/data/model/IngredientModel.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class IngredientModel(
     val name: IngredientTypeModel,
-    val cost: Int,
+    val cost: Double,
     val img: String
 
 )
